courses/permissions.py
from rest_framework import permissions


class AuthorOrReadOnly(permissions.IsAuthenticated):
    def has_object_permission(self, request, view, obj):
        return self.has_permission(request, view) and obj.user == request.user
