requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.6
MarkupSafe==2.1.5
mysql-connector-python==8.0.29
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pycparser==2.21
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
